# removed transpose before _dot_ndarray_coo function