--- a/docs/conduct.rst
+++ b/docs/conduct.rst
@@ -137,7 +137,7 @@ available at
 https://www.contributor-covenant.org/version/2/0/code\_of\_conduct.html.
 
 Community Impact Guidelines were inspired by `Mozilla's code of conduct
-enforcement ladder <https://github.com/mozilla/diversity>`__.
+enforcement ladder <:ghuser:`mozilla/diversity`>`__.
 
 For answers to common questions about this code of conduct, see the FAQ
 at https://www.contributor-covenant.org/faq. Translations are available