Using the dense_sparse tensordot benchmark for 100 runs:
New: 0.09302s +- 0.001313s
Old: 0.186462s +- 0.004102s

Closes #536